24 fix memory leaks (#48)
* todo is already done -> removed it

* investigated a possible fix, added comment to src

* added explicit import for chesstypes.h

* reworked memory allocation for draws

* issue is already fixed -> removed todo

* removed unreachable code (cond. ubyte == -1)

* reworked mem allocation for board / draw to text

* refactored ugly bitshift code

* removed unused variable

* added TODO for chessboard invalidation logic

* added TODOs concerning build / release workflows

* granted explicit data ownership to numpy array

* removed minor compiler warnings

* fixed most ref counters, but added segfault :/

* fixed all issues (tests pass again)
- np.frombuffer caused issue -> didn't use it in the unit test
- missing is_simple_board=True when passing a simple board

+++ chesslib/src/chessboard.c
@@ -331,7 +331,7 @@ uint8_t get_bits_at(const uint8_t data_bytes[], size_t arr_size, int bit_index, 
     /* load data bytes into cache */
     bitOffset = bit_index % 8;
     upper = data_bytes[bit_index / 8];
-    lower = (bit_index / 8 + 1 < arr_size) ? data_bytes[bit_index / 8 + 1] : (uint8_t)0x00;
+    lower = ((size_t)bit_index / 8 + 1 < arr_size) ? data_bytes[bit_index / 8 + 1] : (uint8_t)0x00;
     combined = (uint16_t)((uint16_t)upper << 8) | (lower & 0xFF);
 
     /* cut the desired bits from the combined bytes */
